new rule for y at the beginning of a word followed by a vowel
words that start with yí end up sounding like they start with ee to an Australian ear with no indication of a y as in yes. The y is really to indicate a soft kind of sliding introduction to the word, but would mislead the casual reader into thinking its ye as in ye olden days which is wrong. The y is placed into parenthesis to indicate that it is barely voiced at all, and the following vowel is replaced with its Australian equivalent. For example u is an unstressed vowel after y in pinyin so yu at the beginning of a word becomes uh at the beginning of a word